Add simple mac build script.

+++ scripts/package_www_files.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 # Make sure BRACKETS_WWW_SRC environment variable is set
 if [ "$BRACKETS_WWW_SRC" = "" ]; then
   echo "BRACKETS_WWW_SRC environment variable is not set. Aborting."
@@ -5,8 +7,10 @@ if [ "$BRACKETS_WWW_SRC" = "" ]; then
 fi
 
 # Remove existing www directory contents
-rm -rf xcodebuild/Release/Brackets.app/Contents/www/*
-rmdir xcodebuild/Release/Brackets.app/Contents/www
+if [ -d xcodebuild/Release/Brackets.app/Contents/www ]; then
+  rm -rf xcodebuild/Release/Brackets.app/Contents/www/*
+  rmdir xcodebuild/Release/Brackets.app/Contents/www
+fi
 
 # Make the Brackets.app/Contents/www directory
 mkdir xcodebuild/Release/Brackets.app/Contents/www
